backend/requirements.txt
# FastAPI and Server
fastapi==0.109.0
uvicorn[standard]==0.27.0
python-multipart==0.0.6

# Database - PostgreSQL
sqlalchemy==2.0.25
alembic==1.13.1
asyncpg==0.29.0

# Database - Qdrant Vector DB
qdrant-client==1.7.0

# AI/ML - OpenAI and LlamaIndex
openai==1.10.0
llama-index==0.9.48
llama-index-vector-stores-qdrant==0.1.3

# Authentication
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-dotenv==1.0.0

# Data Validation
pydantic==2.5.3
pydantic-settings==2.1.0
email-validator==2.1.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3
httpx==0.26.0
pytest-cov==4.1.0

# Code Quality
black==24.1.1
ruff==0.1.14
mypy==1.8.0

# Utilities
python-dateutil==2.8.2
